requirements.txt
streamlit
google-genai
python-dotenv
Pillow
firebase-admin
cloudinary